Add alert system for poderators to configure (#788)
As per title. Adds a configurable message to allow poderators to set a message for all users (guests optional).

This adds functionality to future "New Yarn.social version out" update message.

Poderator can set Alert Type, Message, and an option to also alert Guests.

The message is only visible if it has text.

The text is also truncated to a maximum of 150 characters.

Fixes #787

Co-authored-by: Phil S <phil@git.ullarah.com>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/788
Reviewed-by: James Mills <james@mills.io>
Co-authored-by: Phil S. <ullarah@noreply@mills.io>
Co-committed-by: Phil S. <ullarah@noreply@mills.io>