Upload iOS shortcut to post text/image to a Yarn.social instance (#851)
This is a very rudimentary iOS shortcut that can be used to quickly upload text and/or a single image to a Yarn.social image.

This was the most painful thing to develop as the shortcut creator is the most laggiest app in existence.

Before you ask the answer is no I have not implemented a way to add a caption to images.

Download the file and open it up in the Shortcuts app.

cc @prologic

Co-authored-by: Phil S <ullarah@noreply@mills.io>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/851
Co-authored-by: Phil S. <ullarah@noreply@mills.io>
Co-committed-by: Phil S. <ullarah@noreply@mills.io>