feature/output_appended (#864)
ACtually that PR is small (just check the last commit), it has been done above the git branch, so compared to the current master it seems big, but if you merge the git PR, that one will be small.

I just added :
- the context to all parser. (I should have doe it before, but ... )
- a new -f --follow option to timeline. (like tail -f)
- created a new parser follow_parser (I don;t really like the name but it's like it tail -f is named, but if you have suggestion ??)

That parser tick each 5 seconds and append the new messages to the output.

(maybe 5 seconds is too much, we should put a bit more 10s, 20s even 1 minute should be ok here and put it configurable on the future)

To test :
```
make build
./yarnc timeline -f
```

PS: I have reused for followParser the defaultParser.
Co-authored-by: f_dutratineesilva <felipe.dutratineesilva@mindgeek.com>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/864
Reviewed-by: James Mills <james@mills.io>
Co-authored-by: tkanos <tkanos@noreply@mills.io>
Co-committed-by: tkanos <tkanos@noreply@mills.io>