Chinese translation improvement (#938)
hi~ mills and other friends
this merge is all about Chinese translation improvement
1. adjust some inappropriate translation
2. complete almost all untranslated part except pod management
3. create a chinese transaltion guideline, see README at https://git.mills.io/jim/yarn-trans-gl or just read it here.

the things i want you to know is that my friends and i came a really cool Chinese name for yarn and yarn.social but I think yarn.social is a **trademark** and **a symbol of you and many others hardwork**, so i decide to **keep yarn.social not translated** but add the cool Chines name in parentheses after it, hope you'll happy about that!

the current guideline is below:

# 目的 Purpose
Yarn.social网站中文翻译准则，目的是确保网站翻译的一致性
Yarn.social website Chinese translation guidline, the purpose is to keep consistency for the whole translation

# 翻译原则 tranlation principles
1. 信雅达，以信为先
forgive me, it would very offensive to tranlate these principles with my low level, because these principles are set out by one of the greatest chinese translator who's name is Yan Fu(严复)

2. 以推特的用语为主，兼顾微博的用语
mainly used the twitter(which is blocked in mainland china) terms and also considering the weibo(which is more popular in china) terms

# 不翻译的词 Words NOT Translated
1. feed
2. token
3. URL
4. yarn.social 但会在末尾括号加上中文名

# 常用名词对照表 Glossary

| English     | 汉语
| ---         | ---
| yarn        | 言，织言
| yarn.social | yarn.social(织言社交)
| pod         | 节点
| instance    | 节点
| link        | 链接
| follower    | 关注者
| following   | 正在关注
| timeline    | 动态
| twt         | 推文
| post        | 帖子
| profile     | 用户资料
| magic-link-auth | 鉴权魔法链接
| mute        | 屏蔽

* yarn.social是一个商标，要保持这个英文商标，把翻译放到括号里
yarn.social is a trademark, remember that, so keep this English trademark, and put in parentheses.

# 改进 Improvements
欢迎进行讨论提供改进意见
improvements are welcomed

Co-authored-by: Jim Green <jimgreen2013@qq.com>
Co-authored-by: jim <chsing@pm.me>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/938
Reviewed-by: James Mills <james@mills.io>
Co-authored-by: jim <jim@noreply@mills.io>
Co-committed-by: jim <jim@noreply@mills.io>